BF: capture recursive aggregation of changed subds with uptodate metadata to extract top's dataset content info again
Previously, this lead to test failures in #21. Consider a nested
dataset hierarchy with recursively aggregated metadata in top.
One subdataset changes, and its meta data is reaggregated (tops
meta data on sub is out of date but sub has aggregated metadata
we want without another re-extraction, and tops own dataset info
is out of date, because sub has new commits. If we aggregated from
top recursively, there will be a change between tops refcommit
and current state, but only with the added condition, re-extraction
is possible (which will lead to a deletion of tops dataset info
and addition of up to date dataset info)