NF: add schema-based metadata verificator
This commit adds a schema-based verificator for yaml that can be used
to verify the metadata files that are to be submitted to describe
studies that should be archived. The schema is purposefully kept simple
in order to ensure that a least a mininmum an metadata is provided
by the maintainers of the study in question.