RF: improve date handling and error messages
Replaced dateutils with strptime. This removes the
disambiguity in date parsing.
Tightened the date test, i. e. start date cannot be
in the future anymore, and start date has to be given,
if the end date is present.

Added ISO Date format parsing, i. e. YYYY-MM-DD.

Parser errors now include an indication of the offending
token. Fixes for "object not allowed" are proposed.

Keywords can now be either a list of words, or a
single value