IMP: add a preliminary root version detection
A root version detection that determines one root
dataset version that contains a given version of
the dataset. There are two issues:

There might be many more versions that contain
the given version of the dataset. Those might
exist in one branch or in many branches.

The git-call is ad-hoc and should be replaced
with the appropriate datalad git-call-method