ENH: improve meta_extract focus
Meta-extract will not try to extract directories
any longer. That means, it will also not extract
a sub-dataset as part of a file-level extraction.

To extract sub-datasets, perform a dataset-level
extraction on them and either add them to a
super-dataset metadata store with the appropriate
additional keys (inter_dataset_path, root_dataset_id,
root_dataset_version), or aggregate the metadata
into the super dataset.

Fix the description: extraction parameters are
supported.