NF: add a first verison of conduct
Conduct allows processing of information through
pipelines of filters. It parallelizes the filter
processing.

A conductor takes one "provider" and an ordered
list of filters. The provider feeds initial
information into the pipeline, i.e. into the first
filter of the pipeline. If information is available,
the next filter will be invoked. Everything that
is emitted by the last filter is returned as result