NF: add context parameter to meta-extract
meta-extract now interprets a context dictionary,
either given as JSON-string or dictionary. It
reads the datset version from the dictionary, if
present, via the key "dataset_version".

If the version is given in the context, meta-extract
does not have to call dataset.repo.get_hexsha(), which
can be quite expensive.