do not report OK pipeline results without path
If a pipeline element does not contain a path
elements we do not report it. The rationale is
that no change on a file was made if no path
is reported.

This might be, for example, if an extractor
and adder ignore an objects type, e.g. "file",
or "dataset".