signal a proper error if a file is not tracked
This commit improves the description and handling
of the `path`-arguments to `meta-extract`. The
argument might be an absolute path, in which case
it will be converted into a path relative to the
dataset root. If it is not an absolute path,
it will be interpreted verbatim as intra-dataset
file path.

If the file path is not tracked in git/git-annexi,
a ValueError is raised.