test extraction from dataset at symlinked location
This commit adds a test to verify the correct
handling of extraction in datasets that are
initialized with a path that is a symlink. In
addition the tests in test_extract are made
pytest-compatible.

The tests in this commit will fail, the next
commit will fix the reason for failing.