add --file-info parameter
This commit extends the FileInfo dataclass
and derives the AnnexedFileInfo class from it.
The classes hold file-information that is
returned by AnnexRepo.get_content_annexinfo(),
or by GitRepo.status().

It adds a parameter to pass JSON-serialized
FileInfo or AnnexedFileInfo objects to the
extract process via arguments, thus releaving
the necessity to invoce git-annex to
determine file-status.