Reduce clutter in sidebar navigation
This change outsources the toctrees for extractors and MetaLad command
API to dedicated files, and references these files in the index' toctree.
This way, the toctree on the front page has the same content (with slightly
different structure), and the depth of the sidebar navigation toctree
only expands to show all available commands once you click on it.

This makes the documentation structure more alike to other packages
in the datalad ecosystem.

Fixes #312