BF: Raise a ValueError when to-be-extracted files don't exist
Meta-extract cannot work with non-existing files. Thus, whether or not a
given path exists is an imperative check. Ideally, this check should be done
with proper parameter constraints and input validation as done in datalad-next.
As a step towards this, this change makes meta-extract raise a ValueError
when a provided path does not exist.

Any proper parameter validation that would be implemented in the future would
also raise a ValueError such that the pattern introduced here now would be
consistent when better validation is in place

The new behavior is an improvement from the status quo, which threw an error
at the last possible moment: When an internal status() call on the non-existent
path would not return a result. With this change, the problem is caught as
early as currently possible