start with a schema for metadata
This commit introduces a turtle-document that
should grow into a schema that describes the
internal metadata structure of metalad.

The current version is far from complete and
rather naive, it does not yet take a large
number of schemas/ontologies into account.

I did look at the HCLS-Dataset Descriptions
and it seems to me that their `DatasetVersion`
is probably to much geared to a manually
provided version that has been assigned to
a specific code-version. It includes
concepts that are not easily translated in
commit-based versions (which are used in
metalad), e.g. "previousVersion". Therefore
there is, for now, a class called
DatasetInstance, that refers to a specific
commit in a DataLad dataset.

The git-layer is not modelled. That could
be done by
`https://github.com/justin2004/git_to_rdf`
The intention of the layer described here
is to model the metalad specific elements.