use correct `BaseMetadataExtractor`-class
This commit fixes an issue where an incorrect
`BaseMetadataExtractor`-class was used to detect
the generation of a metadata-extractor. That
lead to faulty generation detections.

This commit also fixes an error output that
lead to the output of the class of a class,
i.e. `type`, instead of the class itself.

Some trailing whitespaces are also removed.